Move activites into their individual packages
