+++ app/src/main/AndroidManifest.xml
@@ -9,11 +9,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ProfileActivity" />
-        <activity android:name=".ExchangeActivity" />
-        <activity android:name=".UnlibraryActivity" />
+        <activity android:name=".profile.ProfileActivity" />
+        <activity android:name=".exchange.ExchangeActivity" />
+        <activity android:name=".unlibrary.UnlibraryActivity" />
         <activity android:name=".library.LibraryActivity" />
-        <activity android:name=".LoginActivity" />
+        <activity android:name=".login.LoginActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
