+++ app/src/main/java/com/example/unlibrary/exchange/ExchangeActivity.java
@@ -1,14 +1,16 @@
-package com.example.unlibrary;
+package com.example.unlibrary.exchange;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class LoginActivity extends AppCompatActivity {
+import com.example.unlibrary.R;
+
+public class ExchangeActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
+        setContentView(R.layout.activity_exchange);
     }
 }
