+++ app/src/main/java/com/example/unlibrary/unlibrary/UnlibraryActivity.java
@@ -1,8 +1,10 @@
-package com.example.unlibrary;
+package com.example.unlibrary.unlibrary;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.unlibrary.R;
 
 public class UnlibraryActivity extends AppCompatActivity {
